applied aggrigation to get_data.py
