+++ app/integrators/covid_api_v2_integrator.py
@@ -24,8 +24,7 @@ from models.covid_api_v2_model import (ActiveModel, ConfirmedModel,
                                          TimeseriesUSDataModel,
                                          TimeseriesUSInfoModel,
                                          TimeseriesUSModel, TotalModel)
-from utils.get_data import (DailyReports, get_data_lookup_table,
-                              get_data_time_series, get_US_time_series)
+from utils.get_data import COVIDDataForAPIv2, get_data_lookup_table
 
 
 class CovidAPIv2Integrator:
@@ -36,7 +35,7 @@ class CovidAPIv2Integrator:
             "ts": int = "{timestamp}
         }
     """
-    def __init__(self, daily_reports: DailyReports) -> None:
+    def __init__(self) -> None:
         """ Initiate instances """
         self.lookup_table = get_data_lookup_table()
         self.scheme = {
@@ -44,7 +43,6 @@ class CovidAPIv2Integrator:
             'dt': None,
             'ts': None
         }
-        self.daily_reports=daily_reports
     
     def wrap_data(func) -> ResponseModel:
         """ Wrap a result in a schemed data """
@@ -71,7 +69,7 @@ class CovidAPIv2Integrator:
     def get_current(self) -> List[CurrentModel]:
         """ Current data from all locations (Lastest date) """
         concerned_columns = ['Confirmed', 'Deaths', 'Recovered', 'Active']
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         self.df_grp_by_country = self.df.groupby('Country_Region')[concerned_columns].sum()
         self.df_grp_by_country[concerned_columns] = self.df_grp_by_country[concerned_columns].astype(int)
 
@@ -89,7 +87,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_current_US(self) -> List[CurrentUSModel]:
         """ Get current data for USA's situation """
-        self.df_US = self.daily_reports.get_data_daily_reports(US=True) # Get base data
+        self.df_US = COVIDDataForAPIv2.get_data_daily_reports(US=True) # Get base data
 
         concerned_columns = ['Confirmed', 'Deaths', 'Recovered', 'Active']
         df = self.df_US.groupby(['Province_State'])[concerned_columns].sum().sort_values(by='Confirmed', ascending=False)
@@ -108,7 +106,7 @@ class CovidAPIv2Integrator:
     def get_country(self, country_name: str) -> CountryModel:
         """ Get a country data from its name or ISO 2 """
         concerned_columns = ['Confirmed', 'Deaths', 'Recovered', 'Active']
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         self.df_grp_by_country = self.df.groupby('Country_Region')[concerned_columns].sum()
         self.df_grp_by_country[concerned_columns] = self.df_grp_by_country[concerned_columns].astype(int)
 
@@ -138,7 +136,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_confirmed(self) -> ConfirmedModel:
         """ Summation of all confirmed cases """
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         data = ConfirmedModel(
             confirmed=int(self.df['Confirmed'].sum())
         )
@@ -150,7 +148,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_deaths(self) -> DeathsModel:
         """ Summation of all deaths """
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         data = DeathsModel(
             deaths=int(self.df['Deaths'].sum())
         )
@@ -162,7 +160,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_recovered(self) -> RecoveredModel:
         """ Summation of all recovers """
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         data = RecoveredModel(
             recovered=int(self.df['Recovered'].sum())
         )
@@ -174,7 +172,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_active(self) -> ActiveModel:
         """ Summation of all actives """
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         data = ActiveModel(
             active=int(self.df['Active'].sum())
         )
@@ -186,7 +184,7 @@ class CovidAPIv2Integrator:
     @wrap_data
     def get_total(self) -> TotalModel:
         """ Summation of Confirmed, Deaths, Recovered, Active """
-        self.df = self.daily_reports.get_data_daily_reports() # Get base data
+        self.df = COVIDDataForAPIv2.get_data_daily_reports() # Get base data
         data = TotalModel(
             confirmed=int(self.df['Confirmed'].sum()),
             deaths=int(self.df['Deaths'].sum()),
@@ -204,7 +202,7 @@ class CovidAPIv2Integrator:
             1.) global
             2.) confirmed, deaths, recovered
         """
-        self.df_time_series = get_data_time_series() # Get base data
+        self.df_time_series = COVIDDataForAPIv2.get_data_time_series() # Get base data
 
         if case not in ['global']:
             raw_data = self.df_time_series[case].T.to_dict()
@@ -272,7 +270,7 @@ class CovidAPIv2Integrator:
         if case not in ['confirmed', 'deaths']:
             data = []
         else:
-            self.df_US_time_series = get_US_time_series() # Get base data
+            self.df_US_time_series = COVIDDataForAPIv2.get_US_time_series() # Get base data
             raw_data = self.df_US_time_series[case].T.to_dict()
             data = self.__extract_US_time_series(raw_data)
